feat(CLOUDDEV-455): added new edgecentercloud-go client to provider initialising
